+++ konva.js
@@ -1,8 +1,8 @@
 /*
- * Konva JavaScript Framework v2.4.0
+ * Konva JavaScript Framework v2.4.2
  * http://konvajs.github.io/
  * Licensed under the MIT
- * Date: Fri Sep 21 2018
+ * Date: Tue Oct 16 2018
  *
  * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
  * Modified work Copyright (C) 2014 - present by Anton Lavrenov (Konva)
@@ -21,7 +21,7 @@
 
   var Konva = {
     // public
-    version: '2.4.0',
+    version: '2.4.2',
 
     // private
     stages: [],
@@ -2698,16 +2698,27 @@
      * });
      */
     cache: function(config) {
-      var conf = config || {},
+      var conf = config || {};
+      var rect = {};
+
+      // don't call getClientRect if we have all attributes
+      // it means call it only if have one undefined
+      if (
+        conf.x === undefined ||
+        conf.y === undefined ||
+        conf.width === undefined ||
+        conf.height === undefined
+      ) {
         rect = this.getClientRect({
           skipTransform: true,
           relativeTo: this.getParent()
-        }),
-        width = conf.width || rect.width,
+        });
+      }
+      var width = conf.width || rect.width,
         height = conf.height || rect.height,
         pixelRatio = conf.pixelRatio,
-        x = conf.x || rect.x,
-        y = conf.y || rect.y,
+        x = conf.x === undefined ? rect.x : conf.x,
+        y = conf.y === undefined ? rect.y : conf.y,
         offset = conf.offset || 0,
         drawBorder = conf.drawBorder || false;
 
@@ -3313,15 +3324,20 @@
     isVisible: function() {
       return this._getCache(VISIBLE, this._isVisible);
     },
-    _isVisible: function() {
+    _isVisible: function(relativeTo) {
       var visible = this.getVisible(),
         parent = this.getParent();
 
+      if (relativeTo === parent && visible === 'inherit') {
+        return true;
+      } else if (relativeTo === parent) {
+        return visible;
+      }
       // the following conditions are a simplification of the truth table above.
       // please modify carefully
       if (visible === 'inherit') {
         if (parent) {
-          return parent.isVisible();
+          return parent._isVisible(relativeTo);
         } else {
           return true;
         }
@@ -3336,10 +3352,11 @@
      * @memberof Konva.Node.prototype
      * @returns {Boolean}
      */
-    shouldDrawHit: function(canvas) {
+    shouldDrawHit: function() {
       var layer = this.getLayer();
+
       return (
-        (canvas && canvas.isCache) ||
+        (!layer && this.isListening() && this.isVisible()) ||
         (layer &&
           layer.hitGraphEnabled() &&
           this.isListening() &&
@@ -8404,10 +8421,9 @@
         var rect = child.getClientRect({ relativeTo: that });
 
         // skip invisible children (like empty groups)
-        // or don't skip... hmmm...
-        // if (rect.width === 0 && rect.height === 0) {
-        //     return;
-        // }
+        if (rect.width === 0 && rect.height === 0) {
+          return;
+        }
 
         if (minX === undefined) {
           // initial value for first child
@@ -8428,7 +8444,7 @@
       var hasVisible = false;
       for (var i = 0; i < shapes.length; i++) {
         var shape = shapes[i];
-        if (shape.getVisible()) {
+        if (shape._isVisible(this)) {
           hasVisible = true;
           break;
         }
@@ -11045,7 +11061,10 @@
           }
         } else {
           this._fire(MOUSEUP, { evt: evt, target: this, currentTarget: this });
-          this._fire(CLICK, { evt: evt, target: this, currentTarget: this });
+          if (Konva.listenClickTap) {
+            this._fire(CLICK, { evt: evt, target: this, currentTarget: this });
+          }
+
           if (fireDblClick) {
             this._fire(DBL_CLICK, {
               evt: evt,
@@ -11150,7 +11169,9 @@
         }
       } else {
         this._fire(TOUCHEND, { evt: evt, target: this, currentTarget: this });
-        this._fire(TAP, { evt: evt, target: this, currentTarget: this });
+        if (Konva.listenClickTap) {
+          this._fire(TAP, { evt: evt, target: this, currentTarget: this });
+        }
         if (fireDblClick) {
           this._fire(DBL_TAP, {
             evt: evt,
@@ -15430,6 +15451,7 @@
         maxHeightPx = height - padding * 2,
         currentHeightPx = 0,
         wrap = this.getWrap(),
+        // align = this.getAlign(),
         shouldWrap = wrap !== NONE,
         wrapAtWord = wrap !== CHAR && shouldWrap,
         shouldAddEllipsis = this.getEllipsis() && !shouldWrap;
@@ -15497,6 +15519,9 @@
                   matchWidth = this._getTextWidth(match);
                 }
               }
+              // if (align === 'right') {
+              match = match.trimRight();
+              // }
               this._addTextLine(match);
               textWidth = Math.max(textWidth, matchWidth);
               currentHeightPx += lineHeightPx;
@@ -15511,6 +15536,7 @@
                 break;
               }
               line = line.slice(low);
+              line = line.trimLeft();
               if (line.length > 0) {
                 // Check if the remaining text would fit on one line
                 lineWidth = this._getTextWidth(line);
@@ -16406,6 +16432,9 @@
      * @memberof Konva.Sprite.prototype
      */
     start: function() {
+      if (this.isRunning()) {
+        return;
+      }
       var layer = this.getLayer();
 
       /*
@@ -16798,10 +16827,17 @@
       for (var i = 0; i < points.length / 2; i++) {
         x = points[i * 2];
         y = points[i * 2 + 1];
-        minX = Math.min(minX, x);
-        maxX = Math.max(maxX, x);
-        minY = Math.min(minY, y);
-        maxY = Math.max(maxY, y);
+
+        // skip bad values
+        // TODO: prevent them from parsing function
+        if (!isNaN(x)) {
+          minX = Math.min(minX, x);
+          maxX = Math.max(maxX, x);
+        }
+        if (!isNaN(y)) {
+          minY = Math.min(minY, y);
+          maxY = Math.max(maxY, y);
+        }
       }
       return {
         x: Math.round(minX),
@@ -17132,6 +17168,8 @@
         var parsed = parseFloat(coords[j]);
         if (!isNaN(parsed)) {
           p.push(parsed);
+        } else {
+          p.push(0);
         }
       }
 
@@ -19442,6 +19480,7 @@
   var ATTR_CHANGE_LIST = [
     'resizeEnabledChange',
     'rotateAnchorOffsetChange',
+    'rotateEnabledChange',
     'enabledAnchorsChange',
     'anchorSizeChange',
     'borderEnabledChange',
@@ -19769,7 +19808,6 @@
 
       // add hover styling
       anchor.on('mouseenter', function() {
-        var layer = this.getLayer();
         var tr = this.getParent();
 
         var rad = Konva.getAngle(tr.rotation());
@@ -19779,15 +19817,14 @@
         var isMirrored = scale.y * scale.x < 0;
         var cursor = getCursor(name, rad, isMirrored);
         anchor.getStage().content.style.cursor = cursor;
-        layer.batchDraw();
+        tr._cursorChange = true;
       });
       anchor.on('mouseout', function() {
-        var layer = this.getLayer();
-        if (!layer) {
+        if (!anchor.getStage() || !this.getParent()) {
           return;
         }
         anchor.getStage().content.style.cursor = '';
-        layer.batchDraw();
+        this.getParent()._cursorChange = false;
       });
       this.add(anchor);
     },
@@ -19866,6 +19903,8 @@
       resizerNode.setAbsolutePosition(newAbsPos);
 
       var keepProportion = this.keepRatio() || e.shiftKey;
+
+      // console.log(keepProportion);
 
       if (this.movingResizer === 'top-left') {
         if (keepProportion) {
@@ -20024,6 +20063,8 @@
         var bottomOffsetX = this.getWidth() - bottomRight.x();
         var bottomOffsetY = this.getHeight() - bottomRight.y();
 
+        console.log(topOffsetX, topOffsetY, bottomOffsetX, bottomOffsetY);
+
         bottomRight.move({
           x: -topOffsetX,
           y: -topOffsetY
@@ -20044,6 +20085,8 @@
 
       var height =
         this.findOne('.bottom-right').y() - this.findOne('.top-left').y();
+
+      // console.log(x, y, width, height);
 
       this._fitNodeInto(
         {
@@ -20241,6 +20284,10 @@
       }
     },
     destroy: function() {
+      // console.log(this.isTransforming() && this.getStage());
+      if (this.getStage() && this._cursorChange) {
+        this.getStage().content.style.cursor = '';
+      }
       Konva.Group.prototype.destroy.call(this);
       this.detach();
       this._removeEvents();