+++ src/shapes/Transformer.js
@@ -4,6 +4,7 @@
   var ATTR_CHANGE_LIST = [
     'resizeEnabledChange',
     'rotateAnchorOffsetChange',
+    'rotateEnabledChange',
     'enabledAnchorsChange',
     'anchorSizeChange',
     'borderEnabledChange',
@@ -331,7 +332,6 @@
 
       // add hover styling
       anchor.on('mouseenter', function() {
-        var layer = this.getLayer();
         var tr = this.getParent();
 
         var rad = Konva.getAngle(tr.rotation());
@@ -341,15 +341,14 @@
         var isMirrored = scale.y * scale.x < 0;
         var cursor = getCursor(name, rad, isMirrored);
         anchor.getStage().content.style.cursor = cursor;
-        layer.batchDraw();
+        tr._cursorChange = true;
       });
       anchor.on('mouseout', function() {
-        var layer = this.getLayer();
-        if (!layer) {
+        if (!anchor.getStage() || !this.getParent()) {
           return;
         }
         anchor.getStage().content.style.cursor = '';
-        layer.batchDraw();
+        this.getParent()._cursorChange = false;
       });
       this.add(anchor);
     },
@@ -428,6 +427,8 @@
       resizerNode.setAbsolutePosition(newAbsPos);
 
       var keepProportion = this.keepRatio() || e.shiftKey;
+
+      // console.log(keepProportion);
 
       if (this.movingResizer === 'top-left') {
         if (keepProportion) {
@@ -586,6 +587,8 @@
         var bottomOffsetX = this.getWidth() - bottomRight.x();
         var bottomOffsetY = this.getHeight() - bottomRight.y();
 
+        console.log(topOffsetX, topOffsetY, bottomOffsetX, bottomOffsetY);
+
         bottomRight.move({
           x: -topOffsetX,
           y: -topOffsetY
@@ -606,6 +609,8 @@
 
       var height =
         this.findOne('.bottom-right').y() - this.findOne('.top-left').y();
+
+      // console.log(x, y, width, height);
 
       this._fitNodeInto(
         {
@@ -803,6 +808,10 @@
       }
     },
     destroy: function() {
+      // console.log(this.isTransforming() && this.getStage());
+      if (this.getStage() && this._cursorChange) {
+        this.getStage().content.style.cursor = '';
+      }
       Konva.Group.prototype.destroy.call(this);
       this.detach();
       this._removeEvents();