+++ CHANGELOG.md
@@ -5,17 +5,41 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [new version][unreleased]
 
+### Fixed
+* Performance fixes for caching
+
+## [2.4.2][2018-10-12]
+
+### Fixed
+
+* Fixed a wrong cache when a shape inside group has `listening = false`
+
+## [2.4.1][2018-10-08]
+
+### Changed
+
+* Added some text trim logic to wrap in better
+
+### Fixed
+
+* `getClientRect` for complex paths fixes
+* `getClientRect` calculation fix for groups
+* Update `Konva.Transformer` on `rotateEnabled` change
+* Fix click stage event on dragend
+* Fix some Transformer cursor behavior
+
 ## [2.4.0][2018-09-19]
+
+### Added
+
+* Centered resize with ALT key for `Konva.Transformer`
+* New `centeredScaling` for `Konva.Transformer`
 
 ### Fixed
 
 * Tween support for gradient properties
 * Add `user-select: none` to the stage container to fix some "selected contend around" issues
 
-### Added
-
-* Centered resize with ALT key for `Konva.Transformer`
-* New `centeredScaling` for `Konva.Transformer`
 
 ## [2.3.0][2018-08-30]